mass create domains (url_shortener)

+++ app/Http/Controllers/Api/ShortDomainsApiController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\ApiController;
 use App\Http\Middleware\CheckAdminRole;
+use App\Models\CampaignShortUrl;
 use App\Models\UrlShortener;
 use App\Services\UrlShortener\UrlShortenerService;
 use Illuminate\Http\JsonResponse;
@@ -37,13 +38,19 @@ class ShortDomainsApiController extends ApiController
      */
     public function store(Request $request): JsonResponse
     {
-        $response = $this->urlShortenerService->create($request);
+        $request->validate([
+            'endpoints' => 'required|array',
+            'endpoints.*' => 'required|string|max:180|unique:url_shorteners,name',
+        ]);
+
+        // create new short domain records
+        $response = $this->urlShortenerService->create($request->endpoints);
 
         if (isset($response['error'])) {
             return $this->responseError(message: $response['error'], status:422);
         }
 
-        return $this->responseSuccess($response, 'URL Shortener created successfully.');
+        return $this->responseSuccess(message: $response['message']);
     }
 
     /**
@@ -52,6 +59,12 @@ class ShortDomainsApiController extends ApiController
      */
     public function destroy(int $id): JsonResponse
     {
+        $biddingRecords = CampaignShortUrl::where('url_shortener_id', $id)->first();
+
+        if ($biddingRecords) {
+            return $this->responseError(message: 'URL Shortener cannot be deleted because it is being used in a campaign.', status: 422);
+        }
+
         if (UrlShortener::destroy($id)) {
             return $this->responseSuccess(message: 'URL Shortener deleted successfully.');
         }
